+++ Sources/SwiftQueue/Job.swift
@@ -16,6 +16,11 @@ public final class JobBuilder {
         self.info = JobInfo(type: type)
     }
 
+    /// Get job type
+    public var type: String {
+        return info.type
+    }
+
     /// Allow only 1 job at the time with this ID scheduled or running
     /// Same job scheduled with same id will result in onRemove(TaskAlreadyExist) if override = false
     /// If override = true the previous job will be canceled and the new job will be scheduled
@@ -26,11 +31,21 @@ public final class JobBuilder {
         return self
     }
 
+    /// Get uuid for single instance job
+    public var uuid: String {
+        return info.uuid
+    }
+
     /// Job in different groups can run in parallel
     public func group(name: String) -> Self {
         assertNotEmptyString(name)
         info.group = name
         return self
+    }
+
+    /// Get job group name
+    public var groupName: String {
+        return info.group
     }
 
     /// Delay the execution of the job.
@@ -41,10 +56,20 @@ public final class JobBuilder {
         return self
     }
 
+    /// Get job delay time
+    public var delayTime: TimeInterval {
+        return info.delay ?? 0
+    }
+
     /// Job should be removed from the queue after a certain date
     public func deadline(date: Date) -> Self {
         info.deadline = date
         return self
+    }
+
+    /// Get job dead line
+    public var deadline: Date? {
+        return info.deadline
     }
 
     /// Repeat job a certain number of time and with a interval between each run 
@@ -52,6 +77,16 @@ public final class JobBuilder {
     @available(*, unavailable, message: "Use periodic(Limit, TimeInterval) instead")
     public func periodic(count: Int = -1, interval: TimeInterval = 0) -> Self {
         fatalError("Should not be called")
+    }
+
+    /// Get maximum run count for periodic job
+    public var maxRun: Limit {
+        return info.maxRun
+    }
+
+    /// Get interval for periodic job
+    public var interval: TimeInterval {
+        return info.interval
     }
 
     public func periodic(limit: Limit = .unlimited, interval: TimeInterval = 0) -> Self {
@@ -67,10 +102,20 @@ public final class JobBuilder {
         return self
     }
 
+    /// Get job network requirement
+    public var requireNetwork: NetworkType {
+        return info.requireNetwork
+    }
+
     /// Job should be persisted. 
     public func persist(required: Bool) -> Self {
         info.isPersisted = required
         return self
+    }
+
+    /// Get whether the job should be persisted
+    public var persist: Bool {
+        return info.isPersisted
     }
 
     /// Max number of authorised retry before the job is removed
@@ -86,6 +131,11 @@ public final class JobBuilder {
         return self
     }
 
+    /// Get job max retries
+    public var maxRetries: Limit {
+        return info.retries
+    }
+
     /// Custom tag to mark the job
     public func addTag(tag: String) -> Self {
         assertNotEmptyString(tag)
@@ -93,10 +143,20 @@ public final class JobBuilder {
         return self
     }
 
+    /// Get job tags
+    public var tag: Set<String> {
+        return info.tags
+    }
+
     /// Custom parameters will be forwarded to create method
     public func with(params: [String: Any]) -> Self {
         info.params = params
         return self
+    }
+
+    /// Get job params
+    public var params: [String: Any] {
+        return info.params
     }
 
     internal func build(job: Job) -> SwiftQueueJob {