+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## Feature Branches
+
+### feature/logging
+Added read-only access to JobBuilder properties for centralized debug and logging purpose
+Added func peekQueue() to see what is happening inside the queue such as the running job count and so 
+
 ## [v1.4.1](https://github.com/lucas34/SwiftQueue/tree/1.4.1)
 
 #### Bug fix 
@@ -107,4 +113,4 @@ Support all targets when using carthage
 
 ## [v1.0.0](https://github.com/lucas34/SwiftQueue/tree/1.0.0)
 
-- First stable release
+- First stable release.